+++ Client/Include/UserComponent/Stage_Com.h
@@ -16,7 +16,7 @@ public:
 	void SetPlayer(Player_Com* Player) { m_MainPlayer = Player; }
 	void CreateEatting(const Vector3& Pos, const Vector4& RGB, size_t ID);
 	size_t GetUpdateEattingSize() const { return m_UpdateEatting.size(); }
-	void DeleteEatObject(size_t ID);
+	void DeleteEatObject(int ID);
 	Player_Com* GetMainPlayer() { return m_MainPlayer; }
 
 private:
